feat(divider): create sp-divider from sp-rule
