+++ packages/divider/package.json
@@ -1,36 +1,36 @@
 {
-    "name": "@spectrum-web-components/rule",
+    "name": "@spectrum-web-components/divider",
     "publishConfig": {
         "access": "public"
     },
     "repository": {
         "type": "git",
         "url": "https://github.com/adobe/spectrum-web-components.git",
-        "directory": "packages/rule"
+        "directory": "packages/divider"
     },
     "bugs": {
         "url": "https://github.com/adobe/spectrum-web-components/issues"
     },
-    "homepage": "https://adobe.github.io/spectrum-web-components/components/rule",
+    "homepage": "https://adobe.github.io/spectrum-web-components/components/divider",
     "keywords": [
         "spectrum css",
         "web components",
         "lit-element",
         "lit-html"
     ],
-    "version": "0.5.1",
+    "version": "0.0.1",
     "description": "",
-    "main": "./src/index.js",
-    "module": "./src/index.js",
-    "types": "./src/index.d.ts",
+    "main": "src/index.js",
+    "module": "src/index.js",
     "type": "module",
+    "types": "./src/index.d.ts",
     "exports": {
         ".": "./src/index.js",
         "./src/": "./src/",
         "./custom-elements.json": "./custom-elements.json",
         "./package.json": "./package.json",
-        "./sp-rule": "./sp-rule.js",
-        "./sp-rule.js": "./sp-rule.js"
+        "./sp-divider": "./sp-divider.js",
+        "./sp-divider.js": "./sp-divider.js"
     },
     "files": [
         "custom-elements.json",
@@ -51,11 +51,10 @@
     "author": "",
     "license": "Apache-2.0",
     "devDependencies": {
-        "@spectrum-css/divider": "^1.0.0-beta.3",
-        "@spectrum-web-components/action-button": "^0.2.1"
+        "@spectrum-css/divider": "^1.0.0-beta.3"
     },
     "dependencies": {
-        "@spectrum-web-components/base": "^0.3.1",
+        "@spectrum-web-components/base": "^0.3.0",
         "tslib": "^2.0.0"
     }
 }
