+++ packages/divider/stories/divider.stories.ts
@@ -10,20 +10,17 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { html } from '@open-wc/demoing-storybook';
-import { TemplateResult } from '@spectrum-web-components/base';
+import { html, TemplateResult } from '@spectrum-web-components/base';
 
-import '../sp-rule.js';
+import '../sp-divider.js';
 import './typography-decorator.js';
 import '@spectrum-web-components/action-button/sp-action-button.js';
 import '@spectrum-web-components/icon/sp-icon.js';
-import {
-    AlignLeftIcon,
-    AlignRightIcon,
-} from '@spectrum-web-components/icons-workflow';
+import '@spectrum-web-components/icons-workflow/icons/sp-icon-align-left.js';
+import '@spectrum-web-components/icons-workflow/icons/sp-icon-align-right.js';
 
 export default {
-    title: 'Rule',
+    title: 'Divider',
     decorators: [
         (story: () => TemplateResult): TemplateResult => html`
             <typography-decorator .story=${story()}></typography-decorator>
@@ -34,7 +31,7 @@ export default {
 export const large = (): TemplateResult => {
     return html`
         <h2 class="spectrum-Heading spectrum-Heading--sizeM">Large</h2>
-        <sp-rule size="l"></sp-rule>
+        <sp-divider size="l"></sp-divider>
         <p class="spectrum-Body">Page or Section Titles.</p>
     `;
 };
@@ -42,7 +39,7 @@ export const large = (): TemplateResult => {
 export const medium = (): TemplateResult => {
     return html`
         <h3 class="spectrum-Heading spectrum-Heading--sizeS">Medium</h3>
-        <sp-rule size="m"></sp-rule>
+        <sp-divider size="m"></sp-divider>
         <p class="spectrum-Body">
             Divide subsections, or divide different groups of elements (between
             panels, rails, etc.)
@@ -53,7 +50,7 @@ export const medium = (): TemplateResult => {
 export const small = (): TemplateResult => {
     return html`
         <h4 class="spectrum-Heading spectrum-Heading--sizeXS">Small</h4>
-        <sp-rule size="s"></sp-rule>
+        <sp-divider size="s"></sp-divider>
         <p class="spectrum-Body">
             Divide like-elements (tables, tool groups, elements within a panel,
             etc.)
@@ -67,15 +64,11 @@ export const verticalSmall = (): TemplateResult => {
             style="height: var(--spectrum-global-dimension-size-400, 32px); display: flex;"
         >
             <sp-action-button quiet>
-                <sp-icon slot="icon">
-                    ${AlignLeftIcon()}
-                </sp-icon>
+                <sp-icon-align-left slot="icon"></sp-icon-align-left>
             </sp-action-button>
-            <sp-rule size="s" vertical></sp-rule>
+            <sp-divider vertical size="s"></sp-divider>
             <sp-action-button quiet>
-                <sp-icon slot="icon">
-                    ${AlignRightIcon()}
-                </sp-icon>
+                <sp-icon-align-right slot="icon"></sp-icon-align-right>
             </sp-action-button>
         </div>
     `;
@@ -87,15 +80,27 @@ export const verticalMedium = (): TemplateResult => {
             style="height: var(--spectrum-global-dimension-size-400, 32px); display: flex;"
         >
             <sp-action-button quiet>
-                <sp-icon slot="icon">
-                    ${AlignLeftIcon()}
-                </sp-icon>
+                <sp-icon-align-left slot="icon"></sp-icon-align-left>
             </sp-action-button>
-            <sp-rule size="m" vertical></sp-rule>
+            <sp-divider size="m" vertical></sp-divider>
             <sp-action-button quiet>
-                <sp-icon slot="icon">
-                    ${AlignRightIcon()}
-                </sp-icon>
+                <sp-icon-align-right slot="icon"></sp-icon-align-right>
+            </sp-action-button>
+        </div>
+    `;
+};
+
+export const verticalLarge = (): TemplateResult => {
+    return html`
+        <div
+            style="height: var(--spectrum-global-dimension-size-400, 32px); display: flex;"
+        >
+            <sp-action-button quiet>
+                <sp-icon-align-left slot="icon"></sp-icon-align-left>
+            </sp-action-button>
+            <sp-divider size="l" vertical></sp-divider>
+            <sp-action-button quiet>
+                <sp-icon-align-right slot="icon"></sp-icon-align-right>
             </sp-action-button>
         </div>
     `;
