--- packages/rule/CHANGELOG.md
@@ -1,82 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [0.5.1](https://github.com/adobe/spectrum-web-components/compare/@spectrum-web-components/rule@0.5.0...@spectrum-web-components/rule@0.5.1) (2021-01-28)
-
-**Note:** Version bump only for package @spectrum-web-components/rule
-
-# [0.5.0](https://github.com/adobe/spectrum-web-components/compare/@spectrum-web-components/rule@0.3.3...@spectrum-web-components/rule@0.5.0) (2021-01-21)
-
-### Bug Fixes
-
--   include the "types" entry in package.json files ([b432f59](https://github.com/adobe/spectrum-web-components/commit/b432f5982b3b79f80af12f6d0312cbe2285e608b))
--   stop merging selectors in a way that alters the cascade ([369388f](https://github.com/adobe/spectrum-web-components/commit/369388f8cc147543891087991c569f849ddb9b38))
--   update latest Spectrum CSS beta releases ([d8d3acc](https://github.com/adobe/spectrum-web-components/commit/d8d3acc86de31e58219db6ba2a9d045b83cbe103))
--   use latest @spectrum-css/\* versions ([c35eb86](https://github.com/adobe/spectrum-web-components/commit/c35eb86defd89a0c36b5ea186f6d40f20851b5e5))
-
-### Features
-
--   apply sizedMixin for t-shirt sizing ([d7b63fb](https://github.com/adobe/spectrum-web-components/commit/d7b63fb0db06b5a8a412fea8370964f4db9d18ae))
--   **action-button:** add action button pattern ([03ac00a](https://github.com/adobe/spectrum-web-components/commit/03ac00a710290e6a78340f206d88385a4f8ae8c2))
--   **icons-workflow:** vend fully registered icon components ([941f3a4](https://github.com/adobe/spectrum-web-components/commit/941f3a41486fbd49eca0805fb63383f63313e71e))
--   **rule:** update spectrum css input ([1e9f240](https://github.com/adobe/spectrum-web-components/commit/1e9f240942bfe84370ada81bc8d90b9108011b7f))
-
-# [0.4.0](https://github.com/adobe/spectrum-web-components/compare/@spectrum-web-components/rule@0.3.3...@spectrum-web-components/rule@0.4.0) (2021-01-13)
-
-### Bug Fixes
-
--   include the "types" entry in package.json files ([b432f59](https://github.com/adobe/spectrum-web-components/commit/b432f5982b3b79f80af12f6d0312cbe2285e608b))
--   stop merging selectors in a way that alters the cascade ([369388f](https://github.com/adobe/spectrum-web-components/commit/369388f8cc147543891087991c569f849ddb9b38))
--   update latest Spectrum CSS beta releases ([d8d3acc](https://github.com/adobe/spectrum-web-components/commit/d8d3acc86de31e58219db6ba2a9d045b83cbe103))
--   use latest @spectrum-css/\* versions ([c35eb86](https://github.com/adobe/spectrum-web-components/commit/c35eb86defd89a0c36b5ea186f6d40f20851b5e5))
-
-### Features
-
--   apply sizedMixin for t-shirt sizing ([d7b63fb](https://github.com/adobe/spectrum-web-components/commit/d7b63fb0db06b5a8a412fea8370964f4db9d18ae))
--   **action-button:** add action button pattern ([03ac00a](https://github.com/adobe/spectrum-web-components/commit/03ac00a710290e6a78340f206d88385a4f8ae8c2))
--   **icons-workflow:** vend fully registered icon components ([941f3a4](https://github.com/adobe/spectrum-web-components/commit/941f3a41486fbd49eca0805fb63383f63313e71e))
--   **rule:** update spectrum css input ([1e9f240](https://github.com/adobe/spectrum-web-components/commit/1e9f240942bfe84370ada81bc8d90b9108011b7f))
-
-## [0.3.3](https://github.com/adobe/spectrum-web-components/compare/@spectrum-web-components/rule@0.3.2...@spectrum-web-components/rule@0.3.3) (2020-10-12)
-
-**Note:** Version bump only for package @spectrum-web-components/rule
-
-## [0.3.2](https://github.com/adobe/spectrum-web-components/compare/@spectrum-web-components/rule@0.3.1...@spectrum-web-components/rule@0.3.2) (2020-10-12)
-
-### Bug Fixes
-
--   include default export in the "exports" fields ([f32407d](https://github.com/adobe/spectrum-web-components/commit/f32407d7bbfd18e72c35b6f27740549e79957858))
-
-## [0.3.1](https://github.com/adobe/spectrum-web-components/compare/@spectrum-web-components/rule@0.3.0...@spectrum-web-components/rule@0.3.1) (2020-09-25)
-
-### Bug Fixes
-
--   update side effect listings ([8160d3a](https://github.com/adobe/spectrum-web-components/commit/8160d3ab2c4f5ea11ac40897a5cf1fdaa357f4a8))
-
-# [0.3.0](https://github.com/adobe/spectrum-web-components/compare/@spectrum-web-components/rule@0.2.1...@spectrum-web-components/rule@0.3.0) (2020-08-31)
-
-### Features
-
--   update to Spectrum CSS v3.0.0 ([e8b3d8f](https://github.com/adobe/spectrum-web-components/commit/e8b3d8f75c77c04b4d7af126b91b0f6ad2a40742))
-
-## [0.2.1](https://github.com/adobe/spectrum-web-components/compare/@spectrum-web-components/rule@0.2.0...@spectrum-web-components/rule@0.2.1) (2020-08-19)
-
-**Note:** Version bump only for package @spectrum-web-components/rule
-
-# [0.2.0](https://github.com/adobe/spectrum-web-components/compare/@spectrum-web-components/rule@0.1.1...@spectrum-web-components/rule@0.2.0) (2020-07-17)
-
-### Features
-
--   leverage "exports" field in package.json ([321abd7](https://github.com/adobe/spectrum-web-components/commit/321abd7b7e78ccd9157cff75a1fa3dbd06e81f79))
-
-## [0.1.1](https://github.com/adobe/spectrum-web-components/compare/@spectrum-web-components/rule@0.1.0...@spectrum-web-components/rule@0.1.1) (2020-06-08)
-
-**Note:** Version bump only for package @spectrum-web-components/rule
-
-# 0.1.0 (2020-04-16)
-
-### Features
-
--   **rule:** add Spectrum rule pattern and apply to docs ([f4c52ae](https://github.com/adobe/spectrum-web-components/commit/f4c52ae34eeae80f8edb57828ad93f32f33f713e))
